discordbot/requirements.txt
discord.py==1.6
requests==2.25.0
databases[mysql]==0.4.1
sqlalchemy==1.3
emoji==1.0.1
rich==10.2.2
discord-py-slash-command==3.0.1
